requirements.txt
streamlit
pdfplumber
fpdf
google-generativeai
statistics
